requirements.txt
torch
scipy
monai
nibabel
tqdm
einops
argparse
timm
PyYAML
torchvision
h5py
simpleitk
scikit-learn
imagehash
albumentations
matplotlib
ml_collections
medpy